ajout mean pooling et processus de prediction debugue

+++ src/main/java/fr/univlyon1/configurations/PojoInteraction.java
@@ -33,7 +33,7 @@ public class PojoInteraction<A>{
     private double time;
 
     @XmlElement(name="idObserver")
-    private double idObserver;
+    private long idObserver;
 
     public PojoInteraction(){}
